requirements.txt
regex~=2023.0
aiohttp~=3.0
beautifulsoup4~=4.0
pathvalidate~=2.0
pymorphy3~=2.0
GitPython~=3.0
common[async,sql_async] @ git+https://github.com/Sovamorco/common
discord.py[voice]~=2.0
yoyo-migrations~=8.0
grpcio~=1.0
grpcio-tools~=1.0
pyOpenSSL~=23.0
protobuf~=4.23.4
yt-dlp~=2023.0